+++ src/notion_database/page.py
@@ -4,6 +4,9 @@ from notion_database.children import Children
 from notion_database.properties import Properties
 from notion_database.request import Request
 
+from src.notion_database.cover import Cover
+from src.notion_database.icon import Icon
+
 LOGGER = logging.getLogger("Notion-Database")
 
 
@@ -27,13 +30,15 @@ class Page:
         """
         self.result = self.request.call_api_get(self.url + "/" + page_id)
 
-    def create_page(self, database_id, properties=None, children=None):
+    def create_page(self, database_id, properties=None, children=None, cover: Cover = None, icon: Icon = None):
         """
         Create a page
 
         :param database_id: Identifier for a Notion database
         :param properties: Property values of this page
         :param children: Page content for the new page
+        :param cover:
+        :param icon:
         :return:
         """
         if children is None:
@@ -49,16 +54,22 @@ class Page:
             "properties": properties.result,
             "children": children.result
         }
+        if cover:
+            body.update(cover.result)
+        if icon:
+            body.update(icon.result)
         self.result = self.request.call_api_post(self.url, body)
 
         self.check_field()
 
-    def update_page(self, page_id, properties=None):
+    def update_page(self, page_id, properties=None, cover: Cover = None, icon: Icon = None):
         """
         Update page
 
         :param page_id: Identifier for a Notion page
         :param properties: Property values to update for this page
+        :param cover:
+        :param icon:
         :return:
         """
         if properties is None:
@@ -66,6 +77,10 @@ class Page:
         body = {
             "properties": properties.result,
         }
+        if cover:
+            body.update(cover.result)
+        if icon:
+            body.update(icon.result)
         self.result = self.request.call_api_patch(self.url + "/" + page_id, body)
 
         self.check_field()